Crate aws_config

aws-config provides implementations of region, credential resolution.

These implementations can be used either via the default chain implementation from_env/ConfigLoader or ad-hoc individual credential and region providers.

ConfigLoader can combine different configuration sources into an AWS shared-config: SdkConfig. SdkConfig can be used configure an AWS service client.

Examples

Load default SDK configuration:

let config = aws_config::load_from_env().await;
let client = aws_sdk_dynamodb::Client::new(&config);

Load SDK configuration with a region override:

let region_provider = RegionProviderChain::default_provider().or_else("us-east-1");
let config = aws_config::from_env().region(region_provider).load().await;
let client = aws_sdk_dynamodb::Client::new(&config);

Override configuration after construction of SdkConfig:

let sdk_config = aws_config::load_from_env().await;
let custom_credentials_provider = custom_provider(&sdk_config);
let dynamo_config = aws_sdk_dynamodb::config::Builder::from(&sdk_config)
  .credentials_provider(custom_credentials_provider)
  .build();
let client = aws_sdk_dynamodb::Client::from_conf(dynamo_config);

Modules

connector

Functionality related to creating new HTTP Connectors

credential_process

Credentials Provider for external process

default_provider

Providers that implement the default AWS provider chain Default Provider chains for region, credentials, retries, timeouts and app name.

ecs

Ecs Credentials Provider

environment

Providers that load configuration from environment variables

imds

IMDSv2 Client, credential, and region provider

meta

Meta-providers that augment existing providers with new behavior

profile

Load configuration from AWS Profiles

provider_config

Configuration Options for Credential Providers

sso

SSO Credentials Provider

sts

Credential provider augmentation through the AWS Security Token Service (STS).

timeout

This module defines types that describe timeouts that can be applied to various stages of the Smithy networking stack.

web_identity_token

Load Credentials from Web Identity Tokens

Structs

AppName

App name that can be configured with an AWS SDK client to become part of the user agent string.

ConfigLoader

Load a cross-service SdkConfig from the environment

InvalidAppName

Error for when an app name doesn’t meet character requirements.

RetryConfig

Retry configuration for requests.

Functions

from_env

Create an environment loader for AWS Configuration

load_from_env

Load a default configuration from the environment