Module aws_config::default_provider

Providers that implement the default AWS provider chain Default Provider chains for region, credentials, retries, timeouts and app name.

Typically, this module is used via load_from_env or from_env. It should only be used directly if you need to set custom configuration options to override the default resolution chain.

Modules

app_name

Default app name provider chain

credentials

Default credentials provider chain

region

Default region provider chain

retry_config

Default retry behavior configuration provider chain

timeout_config

Default timeout configuration provider chain