Module aws_config::default_provider::region

Default region provider chain

Typically, this module is used via load_from_env or from_env. It should only be used directly if you need to set custom configuration options to override the default resolution chain.

Structs

Builder

Builder for DefaultRegionChain

DefaultRegionChain

Default region provider chain

Functions

default_provider

Default Region Provider chain