Function aws_config::default_provider::retry_config::default_provider

pub fn default_provider() -> Builder

Default RetryConfig Provider chain

Unlike other “providers” RetryConfig has no related RetryConfigProvider trait. Instead, a builder struct is returned which has a similar API.

This provider will check the following sources in order:

Environment variables

Profile file

Example

use aws_config::default_provider::retry_config;
// Creating a RetryConfig from the default_provider already happens when loading a config from_env
// This is only for illustration purposes
let retry_config = retry_config::default_provider().retry_config().await;
let config = aws_config::from_env().retry_config(retry_config).load().await;
// instantiate a service client:
// <my_aws_service>::Client::new(&config);