Module aws_config::default_provider::timeout_config

Default timeout configuration provider chain

Typically, this module is used via load_from_env or from_env. It should only be used directly if you need to set custom configuration options to override the default resolution chain.

Structs

Builder

Builder for TimeoutConfig that resolves the default timeout configuration

Functions

default_provider

Default TimeoutConfig provider chain