Struct aws_config::identity::LazyCacheBuilder

pub struct LazyCacheBuilder { /* private fields */ }

Builder for lazy identity caching.

Implementations

impl LazyCacheBuilder

pub fn new() -> LazyCacheBuilder

Create a new builder.

pub fn time_source( self, time_source: impl TimeSource + 'static, ) -> LazyCacheBuilder

Set the time source for this cache.

pub fn set_time_source( &mut self, time_source: SharedTimeSource, ) -> &mut LazyCacheBuilder

Set the time source for this cache.

pub fn sleep_impl( self, sleep_impl: impl AsyncSleep + 'static, ) -> LazyCacheBuilder

Set the async sleep implementation for this cache.

pub fn set_sleep_impl( &mut self, sleep_impl: SharedAsyncSleep, ) -> &mut LazyCacheBuilder

Set the async sleep implementation for this cache.

pub fn load_timeout(self, timeout: Duration) -> LazyCacheBuilder

Timeout for identity resolution.

Defaults to 5 seconds.

pub fn set_load_timeout( &mut self, timeout: Option<Duration>, ) -> &mut LazyCacheBuilder

Timeout for identity resolution.

Defaults to 5 seconds.

pub fn buffer_time(self, buffer_time: Duration) -> LazyCacheBuilder

Amount of time before the actual identity expiration time where the identity is considered expired.

For example, if the identity are expiring in 15 minutes, and the buffer time is 10 seconds, then any requests made after 14 minutes and 50 seconds will load a new identity.

Note: random jitter value between [0.0, 0.5] is multiplied to this buffer time.

Defaults to 10 seconds.

pub fn set_buffer_time( &mut self, buffer_time: Option<Duration>, ) -> &mut LazyCacheBuilder

Amount of time before the actual identity expiration time where the identity is considered expired.

For example, if the identity are expiring in 15 minutes, and the buffer time is 10 seconds, then any requests made after 14 minutes and 50 seconds will load a new identity.

Note: random jitter value between [0.0, 0.5] is multiplied to this buffer time.

Defaults to 10 seconds.

pub fn default_expiration(self, duration: Duration) -> LazyCacheBuilder

Default expiration time to set on an identity if it doesn’t have an expiration time.

This is only used if the resolved identity doesn’t have an expiration time set. This must be at least 15 minutes.

Defaults to 15 minutes.

pub fn set_default_expiration( &mut self, duration: Option<Duration>, ) -> &mut LazyCacheBuilder

Default expiration time to set on an identity if it doesn’t have an expiration time.

This is only used if the resolved identity doesn’t have an expiration time set. This must be at least 15 minutes.

Defaults to 15 minutes.

pub fn build(self) -> SharedIdentityCache

Builds a SharedIdentityCache from this builder.

Panics

This builder will panic if required fields are not given, or if given values are not valid.