Enum aws_config::profile::credentials::ProfileFileError

#[non_exhaustive]
pub enum ProfileFileError {
    CouldNotParseProfile(ProfileParseError),
    NoProfilesDefined,
    ProfileDidNotContainCredentials {
        profile: String,
    },
    CredentialLoop {
        profiles: Vec<String>,
        next: String,
    },
    MissingCredentialSource {
        profile: String,
        message: Cow<'static, str>,
    },
    InvalidCredentialSource {
        profile: String,
        message: Cow<'static, str>,
    },
    MissingProfile {
        profile: String,
        message: Cow<'static, str>,
    },
    UnknownProvider {
        name: String,
    },
}

An Error building a Credential source from an AWS Profile

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CouldNotParseProfile(ProfileParseError)

The profile was not a valid AWS profile

NoProfilesDefined

No profiles existed (the profile was empty)

ProfileDidNotContainCredentials

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

profile: String

The name of the profile

The profile did not contain any credential information

CredentialLoop

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

profiles: Vec<String>

Vec of profiles leading to the loop

next: String

The next profile that caused the loop

The profile contained an infinite loop of source_profile references

MissingCredentialSource

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

profile: String

The name of the profile

message: Cow<'static, str>

Error message

The profile was missing a credential source

InvalidCredentialSource

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

profile: String

The name of the profile

message: Cow<'static, str>

Error message

The profile contained an invalid credential source

MissingProfile

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

profile: String

The name of the profile

message: Cow<'static, str>

Error message

The profile referred to a another profile by name that was not defined

UnknownProvider

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

name: String

The name of the provider

The profile referred to credential_source that was not defined