Struct aws_config::profile::credentials::ProfileFileCredentialsProvider

pub struct ProfileFileCredentialsProvider { /* fields omitted */ }

AWS Profile based credentials provider

This credentials provider will load credentials from ~/.aws/config and ~/.aws/credentials. The locations of these files are configurable, see profile::load.

Generally, this will be constructed via the default provider chain, however, it can be manually constructed with the builder:

use aws_config::profile::ProfileFileCredentialsProvider;
let provider = ProfileFileCredentialsProvider::builder().build();

Note: Profile providers to not implement any caching. They will reload and reparse the profile from the file system when called. See lazy_caching for more information about caching.

This provider supports several different credentials formats:

Credentials defined explicitly within the file

[default]
aws_access_key_id = 123
aws_secret_access_key = 456

Assume Role Credentials loaded from a credential source

[default]
role_arn = arn:aws:iam::123456789:role/RoleA
credential_source = Environment

NOTE: Currently only the Environment credential source is supported although it is possible to provide custom sources:

use aws_types::credentials::{self, ProvideCredentials, future};
use aws_config::profile::ProfileFileCredentialsProvider;
#[derive(Debug)]
struct MyCustomProvider;
impl MyCustomProvider {
    async fn load_credentials(&self) -> credentials::Result {
        todo!()
    }
}

impl ProvideCredentials for MyCustomProvider {
  fn provide_credentials<'a>(&'a self) -> future::ProvideCredentials where Self: 'a {
        future::ProvideCredentials::new(self.load_credentials())
    }
}
let provider = ProfileFileCredentialsProvider::builder()
    .with_custom_provider("Custom", MyCustomProvider)
    .build();
}

Assume role credentials from a source profile

[default]
role_arn = arn:aws:iam::123456789:role/RoleA
source_profile = base

[profile base]
aws_access_key_id = 123
aws_secret_access_key = 456

Other more complex configurations are possible, consult test-data/assume-role-tests.json.

Implementations

impl ProfileFileCredentialsProvider

pub fn builder() -> Builder

Builder for this credentials provider