Struct aws_config::provider_config::ProviderConfig

pub struct ProviderConfig { /* private fields */ }

Configuration options for Credential Providers

Most credential providers builders offer a configure method which applies general provider configuration options.

To use a region from the default region provider chain use ProviderConfig::with_default_region. Otherwise, use ProviderConfig::without_region. Note that some credentials providers require a region to be explicitly set.

Implementations

impl ProviderConfig

pub fn without_region() -> Self

Create a default provider config with the region unset.

Using this option means that you may need to set a region manually.

This constructor will use a default value for the HTTPS connector and Sleep implementation when they are enabled as crate features which is usually the correct option. To construct a ProviderConfig without these fields set, use ProviderConfig::empty.

Examples

use aws_config::provider_config::ProviderConfig;
use aws_sdk_sts::Region;
use aws_config::web_identity_token::WebIdentityTokenCredentialsProvider;
let conf = ProviderConfig::without_region().with_region(Some(Region::new("us-east-1")));

let credential_provider = WebIdentityTokenCredentialsProvider::builder().configure(&conf).build();

pub fn empty() -> Self

Constructs a ProviderConfig with no fields set

pub async fn with_default_region() -> Self

Create a default provider config with the region region automatically loaded from the default chain.

Examples

use aws_config::provider_config::ProviderConfig;
use aws_sdk_sts::Region;
use aws_config::web_identity_token::WebIdentityTokenCredentialsProvider;
let conf = ProviderConfig::with_default_region().await;
let credential_provider = WebIdentityTokenCredentialsProvider::builder().configure(&conf).build();
}

pub fn with_region(self, region: Option<Region>) -> Self

Override the region for the configuration

pub async fn load_default_region(self) -> Self

Use the default region chain to set the region for this configuration

Note: the env and fs already set on this provider will be used when loading the default region.

pub fn with_http_connector(self, connector: DynConnector) -> Self

Override the HTTPS connector for this configuration

Warning: Use of this method will prevent you from taking advantage of the HTTP connect timeouts. Consider ProviderConfig::with_tcp_connector.

Stability

This method is expected to change to support HTTP configuration.

pub fn with_tcp_connector<C>(self, connector: C) -> Selfwhere C: Clone + Send + Sync + 'static + Service<Uri>, C::Response: Connection + AsyncRead + AsyncWrite + Send + Unpin + 'static, C::Future: Unpin + Send + 'static, C::Error: Into<Box<dyn Error + Send + Sync + 'static>>,

Override the TCP connector for this configuration

This connector MUST provide an HTTPS encrypted connection.

Stability

This method may change to support HTTP configuration.

pub fn with_sleep(self, sleep: impl AsyncSleep + 'static) -> Self

Override the sleep implementation for this configuration