Struct aws_config::SdkConfig

pub struct SdkConfig { /* private fields */ }

AWS Shared Configuration

Implementations

impl SdkConfig

pub fn region(&self) -> Option<&Region>

Configured region

pub fn endpoint_url(&self) -> Option<&str>

Configured endpoint URL

pub fn retry_config(&self) -> Option<&RetryConfig>

Configured retry config

pub fn timeout_config(&self) -> Option<&TimeoutConfig>

Configured timeout config

pub fn sleep_impl(&self) -> Option<SharedAsyncSleep>

Configured sleep implementation

pub fn identity_cache(&self) -> Option<SharedIdentityCache>

Configured identity cache

pub fn credentials_provider(&self) -> Option<SharedCredentialsProvider>

Configured credentials provider

pub fn token_provider(&self) -> Option<SharedTokenProvider>

Configured bearer auth token provider

pub fn time_source(&self) -> Option<SharedTimeSource>

Configured time source

pub fn app_name(&self) -> Option<&AppName>

Configured app name

pub fn http_client(&self) -> Option<SharedHttpClient>

Configured HTTP client

pub fn use_fips(&self) -> Option<bool>

Use FIPS endpoints

pub fn use_dual_stack(&self) -> Option<bool>

Use dual-stack endpoint

pub fn disable_request_compression(&self) -> Option<bool>

When true, request compression is disabled.

pub fn request_min_compression_size_bytes(&self) -> Option<u32>

Configured minimum request compression size.

pub fn stalled_stream_protection(&self) -> Option<StalledStreamProtectionConfig>

Configured stalled stream protection

pub fn behavior_version(&self) -> Option<BehaviorVersion>

Behavior version configured for this client

pub fn service_config(&self) -> Option<&dyn LoadServiceConfig>

Return an immutable reference to the service config provider configured for this client.

pub fn builder() -> Builder

Config builder

Important: Using the aws-config crate to configure the SDK is preferred to invoking this builder directly. Using this builder directly wonโt pull in any AWS recommended default configuration values.

pub fn to_builder(&self) -> Builder

Convert this SdkConfig into a Builder by cloning it first

pub fn get_origin(&self, setting: &'static str) -> Origin

Get the origin of a setting.

This is used internally to understand how to merge config structs while respecting precedence of origins.

pub fn into_builder(self) -> Builder

Convert this SdkConfig back to a builder to enable modification