Struct aws_config::timeout::Http

pub struct Http { /* private fields */ }

HTTP timeouts used by DynConnectors

Implementations

impl Http

pub fn new() -> Http

Create a new HTTP timeout config with no timeouts set

pub fn read_timeout(&self) -> TriState<Duration>

Return this config’s read timeout.

A limit on the amount of time an application takes to attempt to read the first byte over an established, open connection after write request. This is also known as the “time to first byte” timeout.

pub fn with_read_timeout(self, timeout: TriState<Duration>) -> Http

Mutate this timeout::Http config, setting the HTTP read timeout

pub fn connect_timeout(&self) -> TriState<Duration>

Return this config’s read timeout

A limit on the amount of time after making an initial connect attempt on a socket to complete the connect-handshake.

pub fn with_connect_timeout(self, timeout: TriState<Duration>) -> Http

Mutate this timeout::Http config, setting the HTTP connect timeout

pub fn has_timeouts(&self) -> bool

Return true if any timeouts are intentionally set or disabled

pub fn take_unset_from(self, other: Http) -> Http

Merges two HTTP timeout configs together.