Crate aws_config

aws-config provides implementations of region, credential resolution.

These implementations can be used either via the default chain implementation from_env/ConfigLoader or ad-hoc individual credential and region providers.

ConfigLoader can combine different configuration sources into an AWS shared-config: Config. Config can be used configure an AWS service client.

Examples

Load default SDK configuration:

let config = aws_config::load_from_env().await;
let client = aws_sdk_dynamodb::Client::new(&config);

Load SDK configuration with a region override:

use aws_config::meta::region::RegionProviderChain;
let region_provider = RegionProviderChain::default_provider().or_else("us-east-1");
let config = aws_config::from_env().region(region_provider).load().await;
let client = aws_sdk_dynamodb::Client::new(&config);

Modules

default_provider

Providers that implement the default AWS provider chain Default Provider chains for region and credentials.

environment

Providers that load configuration from environment variables

imds

IMDSv2 Client, credential, and region provider

meta

Meta-providers that augment existing providers with new behavior

profile

Load configuration from AWS Profiles

provider_config

Configuration Options for Credential Providers

sts

Credential provider augmentation through the AWS Security Token Service (STS).

web_identity_token

Load Credentials from Web Identity Tokens

Structs

ConfigLoader

Load a cross-service Config from the environment

Functions

from_env

Create an environment loader for AWS Configuration

load_from_env

Load a default configuration from the environment