aws_config

Module default_provider

Providers that implement the default AWS provider chain

Default Provider chains for region, credentials, retries, timeouts and app name.

Typically, this module is used via load_from_env or from_env. It should only be used directly if you need to set custom configuration options to override the default resolution chain.

Modules

account_id_endpoint_mode

Default provider chain for account-based endpoint mode

app_name

Default app name provider chain

auth_scheme_preference

Default provider chain for auth scheme preference list

checksums

Default provider chains for request/response checksum configuration

credentials

Default credentials provider chain

disable_request_compression

Default “disable request compression” provider chain

endpoint_url

Default endpoint URL provider chain

ignore_configured_endpoint_urls

Default “ignore configured endpoint URLs” provider chain

region

Default region provider chain

request_min_compression_size_bytes

Default “request minimum compression size bytes” provider chain

retry_config

Default retry behavior configuration provider chain

timeout_config

Default timeout configuration provider chain

token

Default access token provider chain

use_dual_stack

Default dual-stack provider chain

use_fips

Default FIPS provider chain