aws_config::profile
Module credentials

Profile File Based Credential Providers

Profile file based providers combine two pieces:

Parsing and resolution of the assume role chain

A user-modifiable hashmap of provider name to provider.

Profile file based providers first determine the chain of providers that will be used to load credentials. After determining and validating this chain, a Vec of providers will be created.

Each subsequent provider will provide boostrap providers to the next provider in order to load the final credentials.

This module contains two sub modules:

repr which contains an abstract representation of a provider chain and the logic to build it from ~/.aws/credentials and ~/.aws/config.

exec which contains a chain representation of providers to implement passing bootstrapped credentials through a series of providers.

Structs

Builder

Builder for ProfileFileCredentialsProvider

ProfileFileCredentialsProvider

AWS Profile based credentials provider

Enums

ProfileFileError

An Error building a Credential source from an AWS Profile