aws_config::retry

Trait ProvideErrorKind

pub trait ProvideErrorKind {
    // Required methods
    fn retryable_error_kind(&self) -> Option<ErrorKind>;
    fn code(&self) -> Option<&str>;
}

Trait that provides an ErrorKind and an error code.

Required Methods

fn retryable_error_kind(&self) -> Option<ErrorKind>

Returns the ErrorKind when the error is modeled as retryable

If the error kind cannot be determined (e.g. the error is unmodeled at the error kind depends on an HTTP status code, return None.

fn code(&self) -> Option<&str>

Returns the code for this error if one exists